+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Release Notes
 
-## [Unreleased](https://github.com/ixocreate/entity-package/compare/0.2.0...develop)
+## [Unreleased](https://github.com/ixocreate/entity-package/compare/0.2.1...develop)
+
+## [v0.2.1 (2019-05-28)](https://github.com/ixocreate/entity-package/compare/0.2.0...0.2.1)
+### Added
+- Entity support for SchemaAwareInterface
+### Fixed
+- Improved EntityCollection performance
 
 ## [v0.2.0 (2019-05-06)](https://github.com/ixocreate/entity-package/compare/0.1.0...0.2.0)
 ### Changed